# src/main/resources/META-INF/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "CreateProductRequest" : {
        "type" : "object",
        "required" : [ "sku", "name", "price" ],
        "description" : "Request object to create a new product",
        "properties" : {
          "sku" : {
            "type" : "string",
            "examples" : [ "SKU123" ],
            "pattern" : "\\S",
            "description" : "Unique SKU of the product"
          },
          "name" : {
            "type" : "string",
            "examples" : [ "Office Chair" ],
            "pattern" : "\\S",
            "description" : "Name of the product"
          },
          "price" : {
            "type" : "number",
            "examples" : [ 50.00 ],
            "description" : "Unit price",
            "minimum" : 0
          },
          "stock" : {
            "type" : "integer",
            "format" : "int32",
            "examples" : [ 10 ],
            "description" : "Initial stock quantity",
            "minimum" : 0
          }
        }
      },
      "Product" : {
        "type" : "object",
        "properties" : {
          "sku" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number"
          },
          "stock" : {
            "type" : "integer",
            "format" : "int32"
          },
          "inStock" : {
            "type" : "boolean"
          }
        }
      },
      "ProductDTO" : {
        "type" : "object",
        "description" : "Product representation for public API responses",
        "properties" : {
          "sku" : {
            "type" : "string",
            "examples" : [ "SKU123" ],
            "description" : "Unique product identifier"
          },
          "name" : {
            "type" : "string",
            "examples" : [ "Office Chair" ],
            "description" : "Product name"
          },
          "price" : {
            "type" : "number",
            "examples" : [ 50.00 ],
            "description" : "Product price"
          },
          "stock" : {
            "type" : "integer",
            "format" : "int32",
            "examples" : [ 10 ],
            "description" : "Available stock quantity"
          }
        }
      },
      "ProductEntity" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sku" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number"
          },
          "stock" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  },
  "tags" : [ {
    "name" : "Product V1",
    "description" : "Basic memory-based API for M1 introduction"
  }, {
    "name" : "Product V2",
    "description" : "Memory-based API with Path Params and Exceptions"
  }, {
    "name" : "Product V3",
    "description" : "Service-based API with Dependency Injection"
  }, {
    "name" : "Product V4",
    "description" : "API using DTOs to encapsulate the Domain model"
  }, {
    "name" : "Product V5",
    "description" : "Validated and documented production-ready API"
  } ],
  "paths" : {
    "/api/v1/products" : {
      "get" : {
        "summary" : "Get all products",
        "description" : "Returns the full list from memory",
        "tags" : [ "Product V1" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create product",
        "description" : "Adds a product if SKU is unique",
        "tags" : [ "Product V1" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Product"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/api/v2/products" : {
      "get" : {
        "summary" : "List all products",
        "tags" : [ "Product V2" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create product",
        "tags" : [ "Product V2" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Product"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "409" : {
            "description" : "SKU already exists"
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/api/v2/products/{sku}" : {
      "get" : {
        "summary" : "Get product by SKU",
        "tags" : [ "Product V2" ],
        "parameters" : [ {
          "description" : "Product SKU",
          "required" : true,
          "name" : "sku",
          "in" : "path",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Product found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Product"
                }
              }
            }
          },
          "404" : {
            "description" : "Product not found"
          }
        }
      },
      "delete" : {
        "summary" : "Delete product",
        "tags" : [ "Product V2" ],
        "parameters" : [ {
          "name" : "sku",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "Not found"
          }
        }
      }
    },
    "/api/v3/products" : {
      "get" : {
        "summary" : "Get all products",
        "tags" : [ "Product V3" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProductEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create product",
        "tags" : [ "Product V3" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProductEntity"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "409" : {
            "description" : "Conflict - SKU already exists"
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/api/v3/products/{sku}" : {
      "get" : {
        "summary" : "Get product by SKU",
        "tags" : [ "Product V3" ],
        "parameters" : [ {
          "description" : "Product SKU",
          "required" : true,
          "name" : "sku",
          "in" : "path",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Product found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductEntity"
                }
              }
            }
          },
          "404" : {
            "description" : "Product not found"
          }
        }
      },
      "delete" : {
        "summary" : "Delete product",
        "tags" : [ "Product V3" ],
        "parameters" : [ {
          "name" : "sku",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "Not found"
          }
        }
      }
    },
    "/api/v4/products" : {
      "get" : {
        "summary" : "Get all products (DTO)",
        "tags" : [ "Product V4" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProductDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create product (DTO)",
        "tags" : [ "Product V4" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "409" : {
            "description" : "SKU already exists"
          }
        }
      }
    },
    "/api/v4/products/{sku}" : {
      "get" : {
        "summary" : "Get product by SKU (DTO)",
        "tags" : [ "Product V4" ],
        "parameters" : [ {
          "description" : "Product SKU",
          "required" : true,
          "name" : "sku",
          "in" : "path",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Product found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Product not found"
          }
        }
      },
      "delete" : {
        "summary" : "Delete product",
        "tags" : [ "Product V4" ],
        "parameters" : [ {
          "name" : "sku",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "Not found"
          }
        }
      }
    },
    "/api/v5/products" : {
      "get" : {
        "summary" : "Get all products",
        "description" : "Returns a list of all products in the catalog",
        "tags" : [ "Product V5" ],
        "responses" : {
          "200" : {
            "description" : "List of products",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProductDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create a new product",
        "description" : "Validates input and persists a new product",
        "tags" : [ "Product V5" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Product created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input or domain validation failure"
          },
          "409" : {
            "description" : "SKU already exists"
          }
        }
      }
    },
    "/api/v5/products/{sku}" : {
      "get" : {
        "summary" : "Get product by SKU",
        "description" : "Finds a specific product using its unique SKU",
        "tags" : [ "Product V5" ],
        "parameters" : [ {
          "description" : "The SKU of the product",
          "examples" : { },
          "required" : true,
          "name" : "sku",
          "in" : "path",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Product found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Product not found"
          }
        }
      },
      "delete" : {
        "summary" : "Delete a product",
        "description" : "Removes a product from the database by its SKU",
        "tags" : [ "Product V5" ],
        "parameters" : [ {
          "description" : "The SKU of the product to delete",
          "required" : true,
          "name" : "sku",
          "in" : "path",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Product deleted"
          },
          "404" : {
            "description" : "Product not found"
          }
        }
      }
    }
  },
  "info" : {
    "title" : "product-catalog API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}